requirements.txt
streamlit>=1.20.0
snowflake-connector-python
pandas
altair